requirements.txt
absl-py==0.9.0
ahrs==0.3.0rc1
aiohttp==3.7.3
alabaster==0.7.12
alembic==1.4.1
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.0
astor==0.8.1
astroid==2.4.2
astropy==4.0.1.post1
async-generator==1.10
async-timeout==3.0.1
atpublic==2.3
attitude-determination==1.0.0
attrs==19.3.0
autopep8==1.5.4
awscli==1.19.87
babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.0
binaryornot==0.4.4
bleach==3.1.5
botocore==1.20.87
cached-property==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.1.1
colorama==0.4.3
comet-git-pure==0.19.16
comet-ml==3.2.0
commonmark==0.9.1
configobj==5.0.6
configparser==5.0.0
contextvars==2.4
cookiecutter==1.7.3
coverage==5.1
cryptography==3.4.7
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
databricks-cli==0.14.3
dataclasses==0.7
dataframe-image==0.1.1
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
dill==0.3.2
diskcache==5.2.1
distro==1.5.0
docker-pycreds==0.4.0
docker==5.0.0
docutils==0.16
dpath==2.0.1
dulwich==0.20.23
dvc==2.3.0
entrypoints==0.3
everett==1.0.2
flake8==3.9.2
flask==2.0.1
flatten-dict==0.4.0
flufl.lock==3.2
focal-loss==0.0.7
fsspec==2021.5.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.2.2
gitdb==4.0.5
gitpython==3.1.7
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-auth==1.30.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gql==0.2.0
grandalf==0.6
graphql-core==1.1
graphviz==0.14.1
greenlet==1.1.0
grpcio==1.29.0
gunicorn==20.1.0
h5py==2.10.0
healpy==1.14.0
httplib2==0.19.1
idna-ssl==1.1.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imagesize==1.2.0
imgaug==0.4.0
immutables==0.15
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.3.4
ipympl==0.7.0
ipython-genutils==0.2.0
ipython==7.16.1
ipywidgets==7.6.3
isort==5.6.4
itsdangerous==2.0.1
jedi==0.17.2
jinja2-time==0.2.0
jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.0
json5==0.9.5
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyter-server==1.8.0
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
jupyterlab==3.0.16
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lineedit==0.1.6
lxml==4.6.2
mailchecker==4.0.7
mako==1.1.4
markdown==3.2.2
markupsafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mlflow==1.17.0
more-itertools==8.4.0
multidict==5.1.0
nanotime==0.5.2
nbclassic==0.3.1
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.9
networkx==2.5.1
nmmn==0.8.9
nodejs==0.1.1
notebook==6.1.3
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opt-einsum==3.2.1
optional-django==0.1.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.1
pathtools==0.1.2
patsy==0.5.1
pdfkit==0.6.1
pexpect==4.8.0
phonenumbers==8.12.24
pickleshare==0.7.5
pillow==7.1.2
pip==20.1.1
pluggy==0.13.1
ply==3.11
poyo==0.5.0
progressbar2==3.51.3
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.2
promise==2.3
prompt-toolkit==3.0.6
protobuf==3.17.2
psutil==5.8.0
ptyprocess==0.6.0
py==1.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.7.0
pycopula==0.1.5
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pydrive2==1.8.3
pyflakes==2.3.1
pygit2==1.6.0
pygments==2.6.1
pygtrie==2.3.2
pylint==2.6.0
pyopenssl==20.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-benedict==0.24.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-fsutil==0.5.0
python-slugify==5.0.2
python-utils==2.4.0
pytz==2020.1
pywavelets==1.1.1
pyyaml==5.3.1
pyzmq==19.0.2
querystring-parser==1.2.4
radian==0.5.10
rchitect==0.3.30
requests-oauthlib==1.3.0
requests==2.23.0
rich==10.2.2
rsa==4.0
ruamel.yaml.clib==0.2.2
ruamel.yaml==0.17.7
s3transfer==0.4.2
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.11.0
send2trash==1.5.0
sentry-sdk==0.16.1
setuptools==49.1.0
shapely==1.7.0
shortuuid==1.0.1
shtab==1.3.6
six==1.14.0
smmap==3.0.4
sniffio==1.2.0
snowballstemmer==2.0.0
soupsieve==2.0
sphinx-rtd-theme==0.5.0
sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlalchemy==1.4.17
sqlparse==0.4.1
src==0.1.0
statsmodels==0.12.1
subprocess32==3.5.4
tabulate==0.8.9
tenning==1.0.6
tensorboard==2.0.2
tensorflow-addons==0.6.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-metadata==0.24.0
tensorflow-probability==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.0.0
tifffile==2020.5.11
toml==0.10.1
tornado==6.1
tqdm==4.46.0
traitlets==4.3.3
transforms3d==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
voluptuous==0.12.1
wandb==0.9.3
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
werkzeug==2.0.1
wheel==0.34.2
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.1
xmltodict==0.12.0
yarl==1.6.3
zc.lockfile==2.0
zipp==3.1.0
